Refactored sftp plugin
